fix slugify and integrate terms api with the new endpoint
