+++ packages/web/services/index.js
@@ -4,4 +4,4 @@ export * from './bookmark';
 export * from './taxonomy';
 export * from './technology';
 export * from './user';
-export * from './term';
+export * from './terms';
